Locomotion feasibility redesign: telemetry, flags, and regression contracts
- Add LocomotionFeasibility, LegContactMode, and height-policy snapshot behind
  LocomotionRedesignConfig / Tuning keys; wire pipeline, stability, body,
  gravity feedforward, replay, and UDP telemetry when enabled.
- Document pathway ownership (LOCOMOTION_KINEMATIC_PATHWAY_AUDIT) and server
  tuning reference; update algorithm docs and test baselines.
- Refactor locomotion regression/navigation acceptance toward pre-fault metrics
  and safer contracts; add test_locomotion_feasibility and related coverage.
- Servo observer, IK/estimator/fusion/safety harness updates; physics sim demo
  tweaks; full verify.sh green.

+++ hexapod-physics-sim/src/demo/serve_mode.cpp
@@ -409,8 +409,10 @@ void ApplyContactArbitrationToLidarSamples(const World& world,
         return;
     }
 
-    const float radius = std::max(0.0f, terrain_config.lidar_contact_arbitration_radius_m);
-    const float disagreement = std::max(0.0f, terrain_config.lidar_contact_disagreement_m);
+    const float radius =
+        std::max(0.0f, static_cast<float>(terrain_config.lidar_contact_arbitration_radius_m));
+    const float disagreement =
+        std::max(0.0f, static_cast<float>(terrain_config.lidar_contact_disagreement_m));
     if (radius <= 1.0e-5f) {
         return;
     }
